requirements.txt
streamlit>=1.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
boto3>=1.34.0
uvicorn>=0.24.0
fastapi>=0.104.0
python-multipart>=0.0.6